Update ITA Localization v. 0.49
Translated the new lines, I used "ToDo" instead of "Todo" to be more appealing in italian, todo looks like a meaningless name in italian (I know it means "to do" in english). What do you think about renaming it in "Slide Over" or something similar to be more Apple consistent? It is quite similar to iPadOS Slide Over function.